refactor(repository): split repo.rs God file into sidecar + per-theme test modules
- Extract filesystem side-effects (image/thumbnail removal, rich-text sidecar
  cleanup, schema-migration / clear directory purge) from repo.rs into a new
  src/repository/sidecar.rs as plain free functions. repo.rs now owns only
  the postcard/redb encode/decode path and delegates filesystem work.
- Move the embedded #[cfg(test)] mod tests out of repo.rs into per-theme
  files under src/repository/tests/{save,pin,dedup,display}_tests.rs so a
  single test edit no longer recompiles the whole 1.7k-line translation unit.
- Replace silently-discarded fs::remove_dir_all errors during schema
  migration / clear with tracing::warn! so orphaned sidecar files are no
  longer dropped without a trace.
- repo.rs shrinks from 1694 to 367 lines; production behavior is unchanged
  and all 411 existing tests still pass.

Refs #70
